Finished Port!
No errors. I'm sure we will need to do some tweaking to get it to work the way we want it to, but all errors have been resolved. I also added the function "String GetConsoleInput()" It returns a string, which can be parsed to integer or double types if need be.